# go/rest/internal/oas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"io"
	"net/url"
	"time"
)

type GetHealthOK struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s GetHealthOK) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

type GetRootOK struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s GetRootOK) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

// Ref: #/components/schemas/HelloResponse
type HelloResponse struct {
	Message   string    `json:"message"`
	Timestamp time.Time `json:"timestamp"`
}

// GetMessage returns the value of Message.
func (s *HelloResponse) GetMessage() string {
	return s.Message
}

// GetTimestamp returns the value of Timestamp.
func (s *HelloResponse) GetTimestamp() time.Time {
	return s.Timestamp
}

// SetMessage sets the value of Message.
func (s *HelloResponse) SetMessage(val string) {
	s.Message = val
}

// SetTimestamp sets the value of Timestamp.
func (s *HelloResponse) SetTimestamp(val time.Time) {
	s.Timestamp = val
}

func (*HelloResponse) v1GetHelloRes() {}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/ProblemDetails
type ProblemDetails struct {
	// サーバによって生成されたHTTPステータスコード.
	Status int32 `json:"status"`
	// - 問題の発生箇所の参照URI
	// - 用途としては、エンドポイントのパスなどを指定する.
	Instance OptString `json:"instance"`
	// 人間が読むことのできるエラー情報の要約（ユーザー/クライアント向け）.
	Title string `json:"title"`
	// 人間が読むことのできるエラー情報の詳細（ユーザー/クライアント向け）
	// - 開発者向けの内部情報は含めない
	// - エラー解消のヒントや入力ミスの理由を提供する.
	Detail OptString `json:"detail"`
	// - 問題の種類を識別するURL
	// -
	// 用途としては、自社/自組織で管理しているエラー詳細ドキュメントへのURLなどを指定する
	// - デフォルトはabout:blank.
	Type url.URL `json:"type"`
}

// GetStatus returns the value of Status.
func (s *ProblemDetails) GetStatus() int32 {
	return s.Status
}

// GetInstance returns the value of Instance.
func (s *ProblemDetails) GetInstance() OptString {
	return s.Instance
}

// GetTitle returns the value of Title.
func (s *ProblemDetails) GetTitle() string {
	return s.Title
}

// GetDetail returns the value of Detail.
func (s *ProblemDetails) GetDetail() OptString {
	return s.Detail
}

// GetType returns the value of Type.
func (s *ProblemDetails) GetType() url.URL {
	return s.Type
}

// SetStatus sets the value of Status.
func (s *ProblemDetails) SetStatus(val int32) {
	s.Status = val
}

// SetInstance sets the value of Instance.
func (s *ProblemDetails) SetInstance(val OptString) {
	s.Instance = val
}

// SetTitle sets the value of Title.
func (s *ProblemDetails) SetTitle(val string) {
	s.Title = val
}

// SetDetail sets the value of Detail.
func (s *ProblemDetails) SetDetail(val OptString) {
	s.Detail = val
}

// SetType sets the value of Type.
func (s *ProblemDetails) SetType(val url.URL) {
	s.Type = val
}

type V1GetHelloBadRequest ProblemDetails

func (*V1GetHelloBadRequest) v1GetHelloRes() {}

type V1GetHelloForbidden ProblemDetails

func (*V1GetHelloForbidden) v1GetHelloRes() {}

type V1GetHelloInternalServerError ProblemDetails

func (*V1GetHelloInternalServerError) v1GetHelloRes() {}

type V1GetHelloUnauthorized ProblemDetails

func (*V1GetHelloUnauthorized) v1GetHelloRes() {}